python-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
qrcode==7.4.2
requests==2.31.0
python-dotenv==1.0.0